Add files via upload
I am thrilled to share the news of my remarkable achievement in successfully clearing the tests for Python, SQL, and Inferential Statistical Analysis. I tackled these challenges with great determination and emerged triumphant. Additionally, I had the privilege of participating in the final round of the Data Analytics and Visualization Challenge organized by Novo Nordisk, which provided me with a golden opportunity to showcase my skills and knowledge in this field